requirements.txt
Django==6.0.4
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
channels==4.3.2
channels_redis==4.3.0
daphne==4.2.1
pillow==12.2.0
